move the source code file to the correct place.
